Add private interface for transforming a database url
This provides a single interface for transforming database urls to be
used in hanami/hanami#1395

Since we agree that there is no need to support customization of this
logic presently, I believe it would be best to provide a single
interface that could be overloaded if someone really needed to change
this.

It also makes unit-testing the behavior simpler.