Support JSON serialization in Hanami::DB::Struct
We would like structs in Hanami to support simple JSON serialization
following the `#to_json` convention.

There have been discussions of this functionality in both ROM and
Dry-Struct projects, where this was rejected as not appropriate for
those gems:

- dry-rb/dry-struct#166
- rom-rb/rom#403

This adds a simple `#to_json` method to `Hanami::DB::Struct` to support
the simple case for serialization.