Use new public API from rom-repository (#16)
`Repository#set_relation` in rom-sql was deprecated, as it was a private API. It didn't set anything so the name was confusing. Now there is a newly added `Repository#prepare_relation`, which is properly public. Use this new method and update the rom gem dependencies to ensure this method is present.